smart: Widen error_cmd reg_feature/reg_count to smallint
ATA Feature and Count registers are 16-bit in 48-bit commands
(smartmontools reports the combined hi/lo bytes), so unsignedTinyInteger
overflows on values >255 (e.g. reg_feature=1344 raised SQLSTATE 22003).
Widen both to unsignedSmallInteger.

+++ database/migrations/2026_06_06_000000_create_smart_tables.php
@@ -151,10 +151,12 @@ return new class extends Migration
             $table->unsignedSmallInteger('error_entry_num');
             $table->unsignedTinyInteger('cmd_slot');
             $table->unsignedTinyInteger('reg_command')->nullable();
-            $table->unsignedTinyInteger('reg_count')->nullable();
+            // Count and Feature are 16-bit in 48-bit ATA commands (smartmontools
+            // reports the combined hi/lo register bytes), so they need smallint.
+            $table->unsignedSmallInteger('reg_count')->nullable();
             $table->unsignedTinyInteger('reg_device')->nullable();
             $table->unsignedTinyInteger('reg_error')->nullable();
-            $table->unsignedTinyInteger('reg_feature')->nullable();
+            $table->unsignedSmallInteger('reg_feature')->nullable();
             $table->unsignedBigInteger('reg_lba')->nullable();
             $table->unsignedInteger('powerup_ms')->nullable();
             $table->string('description', 128)->nullable();
